requirements.txt
line-bot-sdk==3.17.1
flask==3.0.0
python-dotenv>=1.0.1
requests